backend/requirements.txt
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
fastapi==0.111.0
gcloud
google-api-core
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos
grpcio==1.63.0
grpcio-tools
httplib2==0.22.0
idna==3.7
oauth2client==4.1.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
requests==2.31.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.2.1
yandex==0.0.2
yandexcloud==0.278.0
